websearchOpenAI/requirements.txt
streamlit
openai>=1.3.5 
python-dotenv
       